Fix more device info
